LayoutTests/imported/w3c/web-platform-tests/mediacapture-fromelement/HTMLCanvasElement-getImageData-noframe-expected.txt


FAIL CanvasRenderingContext2D.getImageData() does not lead to a frame being captured assert_equals: Video element did not get a frame from the canvas expected 0 but got 4
